Fix min axis array length off by 1 in error message
