+++ packages/lib/src/vis/line/hooks.ts
@@ -1,0 +1,5 @@
+import { createMemo } from '@h5web/shared';
+
+import { getAxisValues } from './utils';
+
+export const useAxisValues = createMemo(getAxisValues);
